create gameController function
After looking at the Connect 4 roject example on TOP, I realized it might be best to create
an all-encompassing gameController function that contains several of the game optics. This
function will house the player array, which stores each player's information. It will also
holdthe activePlayer mechanism that will allow the player's turns to be switched when a move
on the board is made. More functions will soon be contained with this controller. Overall,
this should make it much easier to keep the memory of variables in local storage, thus
reducing the amount of global variables