Get chechBoardWin() functio to work
Successfully got the loop to break once one of the winningCOnditions came back as true
in the 'for ... of' loop. Additionally, the checkBoxes addEventListener is now also
recieving a returned value of true after triggering a run of CheckBoardWin. This true value
can now be used to trigger the end game sequence