Restructure clickBoxes addEventListener to get disabledBoard to function
The entire board now  takes no action when the endGame is reached. Every box on the board
can be clicked but nothing will happen. This was done by separating the forEach loop
that triggers the eventListener from the new clickBoxEvent() function that houses
the actual actions that take place when a box is clicked