Set up Draw check
If all div containers are clicked and no player wins, a message will now be logged
to the console saying that the game has ended in a draw. I still need to replace this message
with a proper callDraw function. This function will log the draw message in the status box
and perhaps even highlight the play again button on the side of the gameboard.