Create turnRead() functionality
The status box will now display which player's turn it is everytime a box is clicked.
More return statements were also added to the boxClickEvent to prevent any more code
being run through if a win or a draw is detected. This keeps the statusbox from
forever switching the player's turns