Fix resetButton functionality
The resetButton isnow working properly. When it is clicked, the gameboard and the scoreboard
is reset. Justa mental note, the biggest of the issues seemed to be trying to run gameController()
a second time in the event Listener. Turns out it was better to leave it alone. The playAgain button
will be developed next