Add Play Again button
Functionality has been given to the playAgainButton. WHen it is clicked, the board, activeplayer,
and statusBox should reset, however the scoreboard remains the same. This allows for players
to play unlimited amounts of rounds and keeptrack of the score. The button's eventListener is housed
within the gameController, as the resetButton is